[ADD] Splitted main into systems and components
