+++ src/Components/Components.hpp
@@ -1,0 +1,16 @@
+/*
+** EPITECH PROJECT, 2022
+** rtype
+** File description:
+** Components
+*/
+
+#ifndef COMPONENTS_HPP_
+#define COMPONENTS_HPP_
+
+#include "Drawable.hpp"
+#include "Controllable.hpp"
+#include "Position.hpp"
+#include "Velocity.hpp"
+
+#endif /* !COMPONENTS_HPP_ */
